agent_project/requirements.txt
google-adk
mysql-connector-python
python-dotenv
